Add resizable_item_pool from DynamicWGPUBufferPool
